grub.d/10_linux: Use an in-house script to generate linux menu entries
Now we have to deal with different kernel configurations for each kernel
variant, and make them to show up in the top-level menu, instead of
showing up inside the "Advanced options for $kernel" submenu.

Also, we want to prefer a kernel with specific configuration over the
other ones, otherwise the machine might not boot, for example, prefer
kernel with 16KiB page size over the one configured with 4KiB page size
for LoongArch64.

* grub.pot: update PO template for new strings in 10_linux

Co-authored-by: Mingcong Bai <jeffbai@aosc.xyz>