10_linux: show the latest kernel for each variant on toplevel menu
Previously, this script just place the latest kernel in the top level
menu, regardless whether other variants are available in this version.

This creates some inconvenience, since there will be only one entry in
the top-level menu, if a newer RC kernel is installed and no other
variants exists for this version.

This major change places the latest kernel for each variants in the
top-level menu, even if the version is older.